Optimizare-Rute - realizare sugestii
Am calculat timpul estimativ, folosind viteza medie de 30 km/h. Am facut metodele private, am eliminat combinatia dintre RutaOptimizata si ViewBag, folosind RutaOptimizataViewModel.

+++ colectare-date/colectare-date/ViewModels/RutaOptimizataViewModel.cs
@@ -6,9 +6,9 @@ namespace colectare_date.ViewModels
     {
         public List<Colectare> Colectari { get; set; } 
         public List<Colectare> ColectariOptimizate { get; set; }
-        public int DistantaTraseuNeoptimizat { get; set; }
-        public int DistantaTraseuOptimizat { get; set; }
-        public int DiferentaDistanta { get; set; }
-        public int TimpEstimativ { get; set; }
+        public double DistantaTraseuNeoptimizat { get; set; }
+        public double DistantaTraseuOptimizat { get; set; }
+        public double DiferentaDistanta { get; set; }
+        public double TimpEstimativ { get; set; }
     }
 }
